Refactor: features/movie-list > features>movies>ui>movie-list 폴더 구조 변경
